deployment/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.16.1
tf_keras==2.16.0
pillow==10.2.0
python-multipart==0.0.9